Reduce downloads of translations via cache-per-day + Disable async-downloading (for now) (#6790)
* ci: Disable (for now) async-downloading of translations

* feat: Allow build-step to re-use already existing translation-files (when NG_DOWNLOAD_TRANSLATIONS_AT_BUILD=false)

* ci: Remove warning about specific retention-length v.s. repo-configured default length

* feat: Always show the Portalicious build/run log directly in the workflow-run log

* refactor: Use consistent `install` pattern. (Should be faster?)

* ci: Remove 'manual override' of Translations-downloading (Always disable in E2E-test runs)

* feat: Cache/re-use downloaded translations (at least once a day / per 'version' of messages.xlf)

+++ interfaces/Portalicious/_build-download-translations.mjs
@@ -1,7 +1,7 @@
 #!/usr/bin/env node
 
 import { LokaliseDownload } from 'lokalise-file-exchange';
-import { accessSync, constants, writeFileSync } from 'node:fs';
+import { accessSync, constants, existsSync, writeFileSync } from 'node:fs';
 
 import { shouldBeEnabled } from './_env.utils.mjs';
 import {
@@ -20,9 +20,15 @@ const requiredTranslations = getRequiredTranslations();
 console.log('Required translations: ', requiredTranslations);
 
 if (!shouldBeEnabled(process.env.NG_DOWNLOAD_TRANSLATIONS_AT_BUILD)) {
-  console.info('Skipping download of translations. Creating mocks instead.');
+  console.info('Skipping download of translations.');
+
   for (const lang of requiredTranslations) {
     const filePath = getTranslationFilePath(lang);
+
+    if (existsSync(filePath)) {
+      console.info(`✅ Translations already available: ${filePath}`);
+      continue;
+    }
 
     writeFileSync(
       filePath,
@@ -57,7 +63,7 @@ await lokaliseDownloader.downloadTranslations({
     original_filenames: false,
   },
   processDownloadFileParams: {
-    asyncDownload: true,
+    asyncDownload: false, // Disabled although recommended by Lokalise. It doesn't work reliable enough.
   },
 });
 console.info(`Download done ✅`);
